Fix warnings detected on cdash ci like conversion de 'size_t' en 'unsigned int', perte possible de données Fix various typos

+++ modules/core/src/math/matrix/vpColVector.cpp
@@ -244,16 +244,16 @@ vpColVector::vpColVector(const vpMatrix &M) : vpArray2D<double>(M.getRows(), 1)
 
 vpColVector::vpColVector(const std::vector<double> &v) : vpArray2D<double>(static_cast<unsigned int>(v.size()), 1)
 {
-  size_t v_size = v.size();
-  for (size_t i = 0; i < v_size; ++i) {
+  unsigned int v_size = static_cast<unsigned int>(v.size());
+  for (unsigned int i = 0; i < v_size; ++i) {
     (*this)[i] = v[i];
   }
 }
 
 vpColVector::vpColVector(const std::vector<float> &v) : vpArray2D<double>(static_cast<unsigned int>(v.size()), 1)
 {
-  size_t v_size = v.size();
-  for (size_t i = 0; i < v_size; ++i) {
+  unsigned int v_size = static_cast<unsigned int>(v.size());
+  for (unsigned int i = 0; i < v_size; ++i) {
     (*this)[i] = static_cast<double>(v[i]);
   }
 }
@@ -350,9 +350,9 @@ vpColVector &vpColVector::operator=(const vpMatrix &M)
 
 vpColVector &vpColVector::operator=(const std::vector<double> &v)
 {
-  size_t v_size = v.size();
+  unsigned int v_size = static_cast<unsigned int>(v.size());
   resize(v_size, false);
-  for (size_t i = 0; i < v_size; ++i) {
+  for (unsigned int i = 0; i < v_size; ++i) {
     (*this)[i] = v[i];
   }
   return *this;
@@ -360,9 +360,9 @@ vpColVector &vpColVector::operator=(const std::vector<double> &v)
 
 vpColVector &vpColVector::operator=(const std::vector<float> &v)
 {
-  size_t v_size = v.size();
+  unsigned int v_size = static_cast<unsigned int>(v.size());
   resize(v_size, false);
-  for (size_t i = 0; i < v_size; ++i) {
+  for (unsigned int i = 0; i < v_size; ++i) {
     (*this)[i] = static_cast<float>(v[i]);
   }
   return *this;
